requirements.txt
langchain==0.3.27
langchain-openai==0.3.30
langchain-groq==0.3.7
langchain-community==0.3.27
beautifulsoup4==4.13.4
selenium==4.35.0
requests-html==0.10.0
slack-sdk==3.36.0
schedule==1.2.2
python-dotenv==1.1.1
